// src/interfaces/common-info.interface.ts
export interface CommonInfoI{
	id:number
	territory: number
	number_of_villages: number
	number_of_smokes: number
	avg_learn: number
	licey: number
	teacher_number: number
	students_number: number
	ogan: number
	chernobyl: number
	hospital: number
	family_medicine:number
	fap: number
	apteka: number
	railroad: number
	airport: number
	district: number
	region: number
	sea: number
}


export interface CommonInfoContainer{
	commonInfo: CommonInfoI
}